+++ src/by/it/group473602/kloster/lesson3/C_HeapMax.java
@@ -38,30 +38,54 @@ import java.util.Scanner;
 public class C_HeapMax {
 
     private class MaxHeap {
-        //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! НАЧАЛО ЗАДАЧИ !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!1
-        //тут запишите ваше решение.
-        //Будет мало? Ну тогда можете его собрать как Generic и/или использовать в варианте B
+
         private List<Long> heap = new ArrayList<>();
 
-        int siftDown(int i) { //просеивание вверх
+        public MaxHeap(){
 
+        }
+        int siftUp(int i) { // просеивание вверх
+            while (heap.get(i) > heap.get((i - 1) / 2)) {
+                swap(i, (i - 1) / 2);
+                i = (i - 1) / 2;
+            }
             return i;
         }
 
-        int siftUp(int i) { //просеивание вниз
-
+        int siftDown(int i) { // просеивание вниз
+            while (2 * i + 1 < heap.size()) {
+                int left = 2 * i + 1; // left
+                int right = 2 * i + 2; // right
+                int j = left;
+                if (right < heap.size() && heap.get(right) < heap.get(left)) {
+                    j = right;
+                }
+                if (heap.get(i) >= heap.get(j))
+                    break;
+                swap(i, j);
+                i = j;
+            }
             return i;
         }
 
-        void insert(Long value) { //вставка
+        void insert(Long value) { // вставка
+            heap.add(value);
+            siftUp(heap.size()-1);
         }
 
-        Long extractMax() { //извлечение и удаление максимума
-            Long result = null;
-
+        Long extractMax() {
+            //Swap max(index=0) element and last element of the tree and then remove max element
+            swap(0, heap.size()-1);
+            Long result = heap.remove(heap.size()-1);
+            siftDown(0);
             return result;
         }
-        //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! КОНЕЦ ЗАДАЧИ !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!1
+
+        private void swap(int first, int second) {
+            Long tmp = heap.get(first);
+            heap.set(first, heap.get(second));
+            heap.set(second, tmp);
+        }
     }
 
     //эта процедура читает данные из файла, ее можно не менять.
@@ -92,7 +116,7 @@ public class C_HeapMax {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson03/heapData.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/kloster/lesson3/heapData.txt");
         C_HeapMax instance = new C_HeapMax();
         System.out.println("MAX="+instance.findMaxValue(stream));
     }
