+++ src/by/it/group473602/kloster/lesson4/A_BinaryFind.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson04;
+package by.it.group473602.kloster.lesson4;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -59,7 +59,7 @@ public class A_BinaryFind {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson04/dataA.txt");
+        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson4/dataA.txt");
         A_BinaryFind instance = new A_BinaryFind();
         //long startTime = System.currentTimeMillis();
         int[] result=instance.findIndex(stream);
