+++ src/by/it/group473602/kloster/lesson4/B_MergeSort.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson04;
+package by.it.group473602.kloster.lesson4;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -82,7 +82,7 @@ public class B_MergeSort {
     }
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson04/dataB.txt");
+        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson4/dataB.txt");
         B_MergeSort instance = new B_MergeSort();
         //long startTime = System.currentTimeMillis();
         int[] result=instance.getMergeSort(stream);
