+++ src/by/it/group473602/kloster/lesson4/C_GetInversions.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson04;
+package by.it.group473602.kloster.lesson4;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -64,7 +64,7 @@ public class C_GetInversions {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson04/dataC.txt");
+        InputStream stream = new FileInputStream(root + "by/it/a_khmelov/lesson4/dataC.txt");
         C_GetInversions instance = new C_GetInversions();
         //long startTime = System.currentTimeMillis();
         int result = instance.calc(stream);
