+++ src/by/it/group473602/kloster/lesson4/dataA.txt
