+++ src/by/it/group473602/kloster/lesson5/A_QSort.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson05;
+package by.it.group473602.kloster.lesson5;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -91,7 +91,7 @@ public class A_QSort {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson05/dataA.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson5/dataA.txt");
         A_QSort instance = new A_QSort();
         int[] result=instance.getAccessory(stream);
         for (int index:result){
