+++ src/by/it/group473602/kloster/lesson5/Lesson5Test.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson05;
+package by.it.group473602.kloster.lesson5;
 
 import org.junit.Test;
 
@@ -12,7 +12,7 @@ public class Lesson5Test {
     @Test
     public void A() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson05/dataA.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson5/dataA.txt");
         A_QSort instance = new A_QSort();
         int[] result=instance.getAccessory(stream);
         boolean ok=Arrays.equals(result,new int[]{1,0,0});
@@ -23,7 +23,7 @@ public class Lesson5Test {
     @Test
     public void B() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson05/dataB.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson5/dataB.txt");
         B_CountSort instance = new B_CountSort();
         int[] result=instance.countSort(stream);
         boolean ok=Arrays.equals(result,new int[]{2,2,3,9,9});
@@ -34,7 +34,7 @@ public class Lesson5Test {
     @Test
     public void C() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson05/dataC.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson5/dataC.txt");
         C_QSortOptimized instance = new C_QSortOptimized();
         int[] result=instance.getAccessory2(stream);
         boolean ok=Arrays.equals(result,new int[]{1,0,0});
