+++ src/by/it/group473602/kloster/lesson5/dataA.txt
