+++ src/by/it/group473602/kloster/lesson6/A_LIS.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson06;
+package by.it.group473602.kloster.lesson6;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -6,39 +6,34 @@ import java.io.InputStream;
 import java.util.Scanner;
 
 /*
-Задача на программирование: наибольшая невозростающая подпоследовательность
+Задача на программирование: наибольшая возростающая подпоследовательность
+см.     https://ru.wikipedia.org/wiki/Задача_поиска_наибольшей_увеличивающейся_подпоследовательности
+        https://en.wikipedia.org/wiki/Longest_increasing_subsequence
 
 Дано:
-    целое число 1<=n<=1E5 ( ОБРАТИТЕ ВНИМАНИЕ НА РАЗМЕРНОСТЬ! )
+    целое число 1≤n≤1000
     массив A[1…n] натуральных чисел, не превосходящих 2E9.
 
 Необходимо:
     Выведите максимальное 1<=k<=n, для которого гарантированно найдётся
     подпоследовательность индексов i[1]<i[2]<…<i[k] <= длины k,
-    для которой каждый элемент A[i[k]] не больше любого предыдущего
-    т.е. для всех 1<=j<k, A[i[j]]>=A[i[j+1]].
-
-    В первой строке выведите её длину k,
-    во второй - её индексы i[1]<i[2]<…<i[k]
-    соблюдая A[i[1]]>=A[i[2]]>= ... >=A[i[n]].
-
-    (индекс начинается с 1)
+    для которой каждый элемент A[i[k]]больше любого предыдущего
+    т.е. для всех 1<=j<k, A[i[j]]<A[i[j+1]].
 
 Решить задачу МЕТОДАМИ ДИНАМИЧЕСКОГО ПРОГРАММИРОВАНИЯ
 
     Sample Input:
     5
-    5 3 4 4 2
+    1 3 3 2 6
 
     Sample Output:
-    4
-    1 3 4 5
+    3
 */
 
+public class A_LIS {
 
-public class C_LongNotUpSubSeq {
 
-    int getNotUpSeqSize(InputStream stream) throws FileNotFoundException {
+    int getSeqSize(InputStream stream) throws FileNotFoundException {
         //подготовка к чтению данных
         Scanner scanner = new Scanner(stream);
         //!!!!!!!!!!!!!!!!!!!!!!!!!     НАЧАЛО ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
@@ -60,10 +55,10 @@ public class C_LongNotUpSubSeq {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson06/dataC.txt");
-        C_LongNotUpSubSeq instance = new C_LongNotUpSubSeq();
-        int result = instance.getNotUpSeqSize(stream);
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson6/dataA.txt");
+        A_LIS instance = new A_LIS();
+        int result = instance.getSeqSize(stream);
         System.out.print(result);
     }
-
 }
+
