+++ src/by/it/group473602/kloster/lesson6/Lesson6Test.java
@@ -1,4 +1,4 @@
-package by.it.group473602.kloster.lesson06;
+package by.it.group473602.kloster.lesson6;
 
 import org.junit.Test;
 
@@ -11,7 +11,7 @@ public class Lesson6Test {
     @Test
     public void A() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson06/dataA.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson6/dataA.txt");
         A_LIS instance = new A_LIS();
         int result=instance.getSeqSize(stream);
         boolean ok=(result==3);
@@ -22,7 +22,7 @@ public class Lesson6Test {
     @Test
     public void B() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson06/dataB.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson6/dataB.txt");
         B_LongDivComSubSeq instance=new B_LongDivComSubSeq();
         int result=instance.getDivSeqSize(stream);
         boolean ok=(result==3);
@@ -32,7 +32,7 @@ public class Lesson6Test {
     @Test(timeout = 1000)
     public void C() throws Exception {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson06/dataC.txt");
+        InputStream stream = new FileInputStream(root + "by/it/group473602/Kloster/lesson6/dataC.txt");
         C_LongNotUpSubSeq instance = new C_LongNotUpSubSeq();
         int result=instance.getNotUpSeqSize(stream);
         boolean ok=(result==4);
