+++ src/by/it/group473602/kloster/lesson6/dataA.txt
